Revert "Introduce es6 chai mocha (#558)"
This reverts commit 0700e7191db54900b64d1bd25bd20bcf23d961ca.
