--- .vscode/jsconfig.json
@@ -1,9 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ES6"
-    },
-    "exclude": [
-        "node_modules",
-        "**/node_modules/*"
-    ]
-}
