--- server/.babelrc
@@ -1,12 +1,0 @@
-{
-  "presets": [
-    "env",
-    "es2015",
-    "stage-0"
-  ],
-  "plugins": [
-    "transform-object-rest-spread"
-  ],
-  "sourceMaps": "inline",
-  "retainLines": true,
-}
