--- server/.build/.gitignore
@@ -1,3 +1,0 @@
-*
-*/
-!.gitignore
