+++ server/package.json
@@ -2,28 +2,16 @@
   "name": "peblio-server",
   "version": "1.0.0",
   "description": "",
-  "main": ".build/server.js",
+  "main": "src/server.js",
   "scripts": {
-    "lint": "./node_modules/.bin/eslint .",
-    "build": "./node_modules/.bin/babel src --out-dir .build --source-maps",
-    "prestartserver": "./node_modules/.bin/babel src --out-dir .build",
-    "startserver": "node .build/server.js",
-    "prebuild": "rm -rf .build/*",
-    "prestart": "npm run build",
-    "start": "AWS_PROFILE=peblio ./scripts/run_with_credstash.sh nodemon .build/server.js",
-    "test": "./node_modules/mocha/bin/mocha --exit --compilers js:babel-core/register test/**/*.js"
+    "start": "AWS_PROFILE=peblio ./scripts/run_with_credstash.sh node_modules/nodemon/bin/nodemon.js",
+    "test": "",
+    "lint": "./node_modules/.bin/eslint ."
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "aws-sdk": "^2.224.1",
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^8.0.4",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
     "bcrypt": "^2.0.1",
     "bcrypt-nodejs": "0.0.3",
     "body-parser": "^1.18.2",
@@ -50,14 +38,8 @@
     "shortid": "^2.2.8"
   },
   "devDependencies": {
-    "babel-preset-react": "^6.24.1",
-    "chai": "^4.2.0",
     "eslint": "^5.3.0",
     "eslint-config-airbnb-base": "^13.0.0",
-    "eslint-plugin-import": "^2.13.0",
-    "mocha": "^5.2.0",
-    "nodemailer-stub-transport": "^1.1.0",
-    "sinon": "^7.2.2",
-    "sinon-chai": "^3.3.0"
+    "eslint-plugin-import": "^2.13.0"
   }
 }
